requirements.txt
gspread>=2.0.0
Telethon==1.18.2
e-objects>=1.2.0
